requirements.txt
fastapi==0.128.4
uvicorn==0.40.0
SQLAlchemy==2.0.46
PyMySQL==1.1.2
lxml==6.0.2
paramiko==4.0.0
APScheduler==3.11.2
python-dotenv==1.2.1